requirements.txt
asgiref==3.7.2
astroid==2.15.6
attrs==23.1.0
black==23.7.0
click==8.1.7
dill==0.3.7
Django==4.2.4
djangorestframework==3.14.0
drf-spectacular==0.26.4
drf-spectacular-sidecar==2023.9.1
flake8==6.1.0
flake8-for-pycharm==0.4.1
inflection==0.5.1
isort==5.12.0
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
lazy-object-proxy==1.9.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
Pillow==10.0.0
platformdirs==3.10.0
pycodestyle==2.11.0
pyflakes==3.1.0
pylint==2.17.5
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.2
rpds-py==0.10.0
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.12.1
typing_extensions==4.7.1
uritemplate==4.1.1
wrapt==1.15.0
